simplify app and remove wagmi and only use viem
