+++ app/evm/page.tsx
@@ -1,14 +1,12 @@
 "use client";
 import { useEffect, useRef, useState } from "react";
-import { ZKPassport, ProofResult, EU_COUNTRIES } from "@zkpassport/sdk";
+import { ZKPassport, ProofResult } from "@zkpassport/sdk";
 import QRCode from "react-qr-code";
-import { useZKPassportVerifier } from "../../lib/useZKPassportVerifier";
-import NetworkInfo from "../../components/NetworkInfo";
+import { createPublicClient, http } from "viem";
+import { sepolia } from "viem/chains";
 
 export default function Home() {
   const [message, setMessage] = useState("");
-  const [firstName, setFirstName] = useState("");
-  const [isEUCitizen, setIsEUCitizen] = useState<boolean | undefined>(undefined);
   const [isOver18, setIsOver18] = useState<boolean | undefined>(undefined);
   const [queryUrl, setQueryUrl] = useState("");
   const [uniqueIdentifier, setUniqueIdentifier] = useState("");
@@ -16,9 +14,6 @@ export default function Home() {
   const [requestInProgress, setRequestInProgress] = useState(false);
   const [onChainVerified, setOnChainVerified] = useState<boolean | undefined>(undefined);
   const zkPassportRef = useRef<ZKPassport | null>(null);
-
-  // Using our custom hook for contract interaction (now using public client)
-  const { verifyProof } = useZKPassportVerifier();
 
   useEffect(() => {
     if (!zkPassportRef.current) {
@@ -30,8 +25,6 @@ export default function Home() {
     if (!zkPassportRef.current) {
       return;
     }
-    setFirstName("");
-    setIsEUCitizen(undefined);
     setMessage("");
     setQueryUrl("");
     setIsOver18(undefined);
@@ -42,8 +35,8 @@ export default function Home() {
     const queryBuilder = await zkPassportRef.current.request({
       name: "ZKPassport",
       logo: "https://zkpassport.id/favicon.png",
-      purpose: "Proof of EU citizenship and firstname",
-      scope: "eu-adult",
+      purpose: "Proof of adulthood",
+      scope: "adult",
       mode: "compressed-evm",
     });
 
@@ -55,12 +48,7 @@ export default function Home() {
       onResult,
       onReject,
       onError,
-    } = queryBuilder
-      .in("nationality", EU_COUNTRIES)
-      .disclose("firstname")
-      .gte("age", 18)
-      .disclose("document_type")
-      .done();
+    } = queryBuilder.gte("age", 18).done();
 
     setQueryUrl(url);
     console.log(url);
@@ -79,29 +67,51 @@ export default function Home() {
 
     const proofs: ProofResult[] = [];
 
-    onProofGenerated((result: ProofResult) => {
+    onProofGenerated(async (result: ProofResult) => {
       console.log("Proof result", result);
       proofs.push(result);
       setMessage(`Proofs received`);
       setRequestInProgress(false);
 
+      if (!zkPassportRef.current) {
+        return;
+      }
+
       try {
         // Perform on-chain verification
-        verifyProof(result)
-          .then(({ isVerified, uniqueIdentifier }) => {
-            setOnChainVerified(isVerified);
-            console.log("Unique identifier (from on-chain verification)", uniqueIdentifier);
-            setMessage(
-              isVerified
-                ? "Proof verified on-chain successfully"
-                : "Proof verification on-chain failed"
-            );
-          })
-          .catch((error) => {
-            console.error("Error verifying on-chain:", error);
-            setMessage("Error during on-chain verification");
-            setOnChainVerified(false);
-          });
+        const params = zkPassportRef.current.getSolidityVerifierParameters(result);
+        const verifierDetails =
+          zkPassportRef.current.getSolidityVerifierDetails("ethereum_sepolia");
+
+        // Create a public client for the selected network
+        const publicClient = createPublicClient({
+          chain: sepolia,
+          transport: http("https://ethereum-sepolia-rpc.publicnode.com"),
+        });
+
+        // Use the public client to call the view function
+        const contractCallResult = await publicClient.readContract({
+          address: verifierDetails.address as `0x${string}`,
+          abi: verifierDetails.abi,
+          functionName: "verifyProof",
+          args: [
+            params.vkeyHash,
+            params.proof,
+            params.publicInputs,
+            params.committedInputs,
+            params.committedInputCounts,
+            params.validityPeriodInDays,
+          ],
+        });
+
+        console.log("Contract call result", contractCallResult);
+        const isVerified = Array.isArray(contractCallResult)
+          ? Boolean(contractCallResult[0])
+          : false;
+        const uniqueIdentifier = Array.isArray(contractCallResult)
+          ? String(contractCallResult[1])
+          : "";
+        setOnChainVerified(isVerified);
       } catch (error) {
         console.error("Error preparing verification:", error);
       }
@@ -110,8 +120,6 @@ export default function Home() {
     onResult(async ({ result, uniqueIdentifier, verified, queryResultErrors }) => {
       console.log("Result of the query", result);
       console.log("Query result errors", queryResultErrors);
-      setFirstName(result?.firstname?.disclose?.result);
-      setIsEUCitizen(result?.nationality?.in?.result);
       setIsOver18(result?.age?.gte?.result);
       setMessage("Result received");
       setUniqueIdentifier(uniqueIdentifier || "");
@@ -144,20 +152,8 @@ export default function Home() {
 
   return (
     <main className="w-full h-full flex flex-col items-center p-10">
-      <NetworkInfo />
-
       {queryUrl && <QRCode className="mb-4" value={queryUrl} />}
       {message && <p>{message}</p>}
-      {firstName && (
-        <p className="mt-2">
-          <b>Firstname:</b> {firstName}
-        </p>
-      )}
-      {typeof isEUCitizen === "boolean" && (
-        <p className="mt-2">
-          <b>Is EU citizen:</b> {isEUCitizen ? "Yes" : "No"}
-        </p>
-      )}
       {typeof isOver18 === "boolean" && (
         <p className="mt-2">
           <b>Is over 18:</b> {isOver18 ? "Yes" : "No"}
