+++ package.json
@@ -10,14 +10,12 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@tanstack/react-query": "^5.74.3",
     "@zkpassport/sdk": "^0.3.0",
     "next": "15.1.0",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-qr-code": "^2.0.15",
-    "viem": "^2.27.0",
-    "wagmi": "^2.14.16"
+    "viem": "^2.27.0"
   },
   "devDependencies": {
     "@types/node": "^20",
